+++ addOns/authhelper/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Maintenance changes.
 - Dropped "to Clipboard" from ZAP copy menu items or buttons (Issue 8179).
+- Update cookies in header based session management, to cope with apps that set them via JavaScript.
 
 ## [0.10.0] - 2023-10-12
 ### Changed